refactor(component): 컴포넌트 폴더 구조 변경
- common: 재사용 가능한 일반적인 컴포넌트
- restaurantListPage: 리스트 관련 컴포넌트
- bottomSheet: 바텀시트 관련 컴포넌트
